(MODULES-7398) PDK Convert the module
This commit converts the module to PDK format.  There are no code or test
changes as part of this conversion.  Note that rubocop testing is effectively
disabled and will be rectified in further pull requests.
