+++ spec/spec_helper.rb
@@ -1,40 +1,45 @@
-require 'pry' if Bundler.rubygems.find_name('pry').any?
+
 require 'puppetlabs_spec_helper/module_spec_helper'
-require 'rake'
-require 'fileutils'
+require 'rspec-puppet-facts'
+
+begin
+  require 'spec_helper_local' if File.file?(File.join(File.dirname(__FILE__), 'spec_helper_local.rb'))
+rescue LoadError => loaderror
+  warn "Could not require spec_helper_local: #{loaderror.message}"
+end
+
+include RspecPuppetFacts
+
+default_facts = {
+  puppetversion: Puppet.version,
+  facterversion: Facter.version,
+}
+
+default_facts_path = File.expand_path(File.join(File.dirname(__FILE__), 'default_facts.yml'))
+default_module_facts_path = File.expand_path(File.join(File.dirname(__FILE__), 'default_module_facts.yml'))
+
+if File.exist?(default_facts_path) && File.readable?(default_facts_path)
+  default_facts.merge!(YAML.safe_load(File.read(default_facts_path)))
+end
+
+if File.exist?(default_module_facts_path) && File.readable?(default_module_facts_path)
+  default_facts.merge!(YAML.safe_load(File.read(default_module_facts_path)))
+end
 
 RSpec.configure do |c|
-  # set the environment variable before files are loaded, otherwise it is too late
-  ENV['ChocolateyInstall'] = 'c:\blah'
-
-  begin
-    Win32::Registry.any_instance.stubs(:[]).with('Bind')
-    Win32::Registry.any_instance.stubs(:[]).with('Domain')
-    Win32::Registry.any_instance.stubs(:[]).with('ChocolateyInstall').raises(Win32::Registry::Error.new(2), 'file not found yo')
-  rescue
-    # we don't care
-  end
-
-  # https://www.relishapp.com/rspec/rspec-core/v/2-12/docs/mock-framework-integration/mock-with-mocha!
-  c.mock_framework = :mocha
-  # see output for all failures
-  c.fail_fast = false
-  c.expect_with :rspec do |e|
-    e.syntax = [:should, :expect]
-  end
-  c.raise_errors_for_deprecations!
-
-  c.after :suite do
-    #result = RubyProf.stop
-    # Print a flat profile to text
-    #printer = RubyProf::FlatPrinter.new(result)
-    #printer.print(STDOUT)
+  c.default_facts = default_facts
+  c.before :each do
+    # set to strictest setting for testing
+    # by default Puppet runs at warning level
+    Puppet.settings[:strict] = :warning
   end
 end
 
-# We need this because the RAL uses 'should' as a method.  This
-# allows us the same behaviour but with a different method name.
-class Object
-  alias :must :should
-  alias :must_not :should_not
+def ensure_module_defined(module_name)
+  module_name.split('::').reduce(Object) do |last_module, next_module|
+    last_module.const_set(next_module, Module.new) unless last_module.const_defined?(next_module)
+    last_module.const_get(next_module)
+  end
 end
+
+# 'spec_overrides' from sync.yml will appear below this line
